Github Issue #62: Address Endianness Issue with OSM Metatiles
It seems Metatiles are endian dependant;
 so as the example test file is little endian style,
 disable tests on big endian machines.

+++ test/check_metatile.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 # Copyright: CC0
+if [ $(echo -n I | od -to2 | awk '{ print substr($2,6,1); exit}') != 1 ]; then
+    echo "Skipping metatile test on big endian machine"
+    echo " sample metatile file is only for little endian machines"
+    exit 0
+fi
+
 if [ -n "$srcdir" ]; then
   ./test_metatile "$srcdir/metatile_example" && rm tilefrommeta.png
 else
